ref: rename logging to tracing
